RSIs_V3
We did the run on Nada’s PC, which has an NVIDIA GeForce RTX 3060, and alhamdulillah, we successfully obtained the results. All outputs — including models, logs, and checkpoints — are saved in the work_dirs folder. The notebook containing Nada’s run is RSIs_Using_VFNet_Fine_tuning_Nada_run.ipynb.

I also tried running our code on Google Colab, but it was not practical due to session instability and limited resources. Therefore, we decided to perform the fine-tuning locally on Nada’s machine instead.

The fine-tuning process took several hours, and for that reason, any future runs will most likely be executed locally on her PC. She will send me the resulting outputs, and I will handle uploading them.

We initially tried to manage this workflow through GitHub, with Nada pushing updates directly to the repository, but we encountered some issues. For now, we’re following this simplified approach. In the future, it might be better if Nada forks the repository and pushes her updates from there.